Added solution for viewport reset #27073 bug.
