+++ .storybook/main.mjs
@@ -1,26 +1,32 @@
-module.exports = {
-  "stories": ["../src/**/*.mdx", "../src/**/*.stories.@(js|jsx|ts|tsx)"],
+import { mergeConfig } from 'vite';
+import vue from '@vitejs/plugin-vue';
 
-  "addons": ["@storybook/addon-links", {
-    name: "@storybook/addon-essentials",
-    options: {
-      actions: false
-    }
-  }, "@storybook/preset-scss", "@whitespace/storybook-addon-html", "@storybook/addon-mdx-gfm"],
-
-  "framework": {
-    name: "@storybook/vue3-vite",
-    options: {}
+const config = {
+  stories: ['../src/**/*.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)'],
+  addons: [
+    '@storybook/addon-links',
+    {
+      name: '@storybook/addon-essentials',
+      options: {
+        actions: false,
+      },
+    },
+    '@storybook/preset-scss',
+    '@whitespace/storybook-addon-html',
+  ],
+  framework: {
+    name: '@storybook/vue3-vite',
+    options: {},
   },
-
-  async viteFinal(config, {
-    configType
-  }) {
-    config.base = process.env.BASE_URL || config.base;
-
-    // return the customized config
-    return config;
+  async viteFinal(config, { configType }) {
+    return mergeConfig(config, {
+      plugins: [vue()],
+      base: process.env.BASE_URL || config.base,
+    });
   },
-
-  docs: {}
+  docs: {
+    autodocs: true,
+  },
 };
+
+export default config;
