+++ package.json
@@ -39,8 +39,6 @@
     "@storybook/addon-actions": "^8.3.3",
     "@storybook/addon-essentials": "^8.3.3",
     "@storybook/addon-links": "^8.3.3",
-    "@storybook/addon-mdx-gfm": "^8.3.3",
-    "@storybook/addons": "^7.6.17",
     "@storybook/preset-scss": "^1.0.3",
     "@storybook/theming": "^8.3.3",
     "@storybook/vue3": "^8.3.3",
